full-stack-web-developer/catalog/requirements.txt

SQLAlchemy
SQLAlchemy-Migrate
WTForm
Flask
Flask-Login
Flask-OpenId
GitHub-Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Bootstrap
Marshmallow